Use qualified name
