+++ dev/services/wms/ows_refactored/sea_ocean_coast/intertidal_c3/style_intertidal_cfg.py
@@ -1,32 +1,3 @@
-from datacube_ows.band_utils import scalable
-
-
-@scalable
-def elevation_adaptive(data, band, lot, hot, band_mapper=None):
-    """
-    Experimental adaptive elevation function, using pixel-level
-    tide metadata to calculate relative elevation for any
-    given location.
-
-    This implementation should be free of any tile-based
-    discontinuities in the resulting visualisation.
-
-    # TODO: Add hillshading
-    """
-
-    # Calculate observed tide range (max - min)
-    otr = data[hot] - data[lot]
-
-    # Calculate distance between elevation and minumum
-    # observed tide height
-    distance_to_min = data[band] - data[lot]
-
-    # Calculate proportion along observed tide range
-    proportion_array = distance_to_min / otr
-
-    return proportion_array
-
-
 legend_intertidal_percentage_by_20 = {
     "begin": "0.0",
     "end": "100",
@@ -204,7 +175,7 @@ style_intertidal_elevation_adaptive = {
     "title": "Elevation",
     "abstract": "Intertidal elevation in metres above Mean Sea Level",
     "index_function": {
-        "function": elevation_adaptive,
+        "function": "ows_refactored.sea_ocean_coast.intertidal_c3.utils_intertidal.elevation_adaptive",
         "mapped_bands": True,
         "kwargs": {
             "band": "elevation",
